refactor: move the measure-unit block into its own module
